Implemented refresh credentials.

+++ strava_flow/strava_api/credentials.py
@@ -8,7 +8,15 @@ from strava_flow.utils.time import datetime_from_timestamp
 from strava_flow.strava_api.oauth2 import OAuth2AuthenticationClient
 
 
+class InvalidTokenException(Exception):
+    """Received token does not contain the valid format"""
+
+    pass
+
+
 class Credentials:
+    _EXPIRES_SOON_OFFSET = 3600.0
+
     def __init__(
         self,
         client_id: int,
@@ -33,10 +41,16 @@ class Credentials:
         self.invalid = True
 
     def access_token_expired(self) -> bool:
+        return self._is_expired()
+
+    def access_token_soon_expired(self) -> bool:
+        return self._is_expired(self._EXPIRES_SOON_OFFSET)
+
+    def _is_expired(self, offset: float = 0.0) -> bool:
         now = datetime.now(timezone.utc)
         if not self.token_expiry:
             return False
-        elif not self.invalid and now < datetime_from_timestamp(self.token_expiry):
+        elif not self.invalid and now < datetime_from_timestamp(self.token_expiry - offset):
             return False
         else:
             return True
@@ -50,7 +64,7 @@ class Credentials:
             access_token=data['access_token'],
             refresh_token=data['refresh_token'],
             token_expiry=data['token_expiry'],
-            scopes=data['scopes'],
+            scopes=data['scope'],
             user_agent=data['user_agent'],
             invalid=data['invalid'],
         )
@@ -85,13 +99,15 @@ class CredentialsStorage:
 
         return credentials
 
-    def put(self, credentials: Credentials) -> None:
-        Path(self._filepath).touch(exist_ok=True)
-        with open(self._filepath, 'w') as f:
-            f.write(credentials.to_json())
+    def put(self, credentials: Optional[Credentials]) -> None:
+        if credentials is not None:
+            Path(self._filepath).touch(exist_ok=True)
+            with open(self._filepath, 'w') as f:
+                f.write(credentials.to_json())
 
     def delete(self) -> None:
-        os.remove(self._filepath)
+        if os.path.exists(self._filepath):
+            os.remove(self._filepath)
 
 
 class StravaCredentialsService:
@@ -108,25 +124,7 @@ class StravaCredentialsService:
         self._client_secret = client_secret
         self._storage = self._initialize_storage()
 
-    def get_credentials(self) -> Optional[Credentials]:
-        credentials = self._storage.get()
-        if not credentials or credentials.invalid or credentials.access_token_expired():
-            credentials = self._get_new_credentials()
-        self._storage.put(credentials)
-        return credentials
-
-    def refresh_credentials(self) -> None:
-        # @todo implement
-        raise NotImplementedError
-
-    def _initialize_storage(self) -> CredentialsStorage:
-        home = str(Path.home())
-        filename = f'{self._client_id}.{self._CREDENTIALS_FILENAME}'
-        filepath = os.path.join(home, self._CREDENTIALS_FOLDER, filename)
-        return CredentialsStorage(filepath)
-
-    def _get_new_credentials(self) -> Credentials:
-        client = OAuth2AuthenticationClient(
+        self._auth_client = OAuth2AuthenticationClient(
             client_id=self._client_id,
             client_secret=self._client_secret,
             scopes=self._SCOPES,
@@ -135,20 +133,54 @@ class StravaCredentialsService:
             token_uri=self._TOKEN_URI,
             revoke_uri=self._REVOKE_URI,
         )
-        token = client.authenticate()
-        # @todo validate token content
+
+    def get_credentials(self) -> Optional[Credentials]:
+        credentials = self._storage.get()
+        if not credentials or credentials.invalid or credentials.access_token_expired():
+            credentials = self._get_new_credentials()
+        self._storage.put(credentials)
+        return credentials
+
+    def refresh_credentials(self) -> Optional[Credentials]:
+        credentials = self._storage.get()
+        if not credentials or credentials.invalid:
+            credentials = self._get_new_credentials()
+        elif credentials.access_token_soon_expired():
+            credentials = self._refresh_existing_credentials(credentials)
+        self._storage.put(credentials)
+        return credentials
+
+    def _initialize_storage(self) -> CredentialsStorage:
+        home = str(Path.home())
+        filename = f'{self._client_id}.{self._CREDENTIALS_FILENAME}'
+        filepath = os.path.join(home, self._CREDENTIALS_FOLDER, filename)
+        return CredentialsStorage(filepath)
+
+    def _get_new_credentials(self) -> Credentials:
+        token = self._auth_client.authenticate()
+        return self._convert_token_to_credentials(token)
+
+    def _refresh_existing_credentials(self, credentials: Credentials) -> Credentials:
+        token = self._auth_client.refresh(credentials.refresh_token)
         return self._convert_token_to_credentials(token)
 
     def _convert_token_to_credentials(self, token_dict: Dict[str, Any]) -> Credentials:
-        return Credentials(
-            client_id=self._client_id,
-            client_secret=self._client_secret,
-            access_token=token_dict['access_token'],
-            refresh_token=token_dict['refresh_token'],
-            token_expiry=token_dict['token_expiry'],
-            scopes=self._SCOPES,
-            user_agent=self._USER_AGENT,
-        )
+        if self._is_token_valid(token_dict):
+            return Credentials(
+                client_id=self._client_id,
+                client_secret=self._client_secret,
+                access_token=token_dict['access_token'],
+                refresh_token=token_dict['refresh_token'],
+                token_expiry=token_dict['token_expiry'],
+                scopes=self._SCOPES,
+                user_agent=self._USER_AGENT,
+            )
+        else:
+            raise InvalidTokenException
+
+    @staticmethod
+    def _is_token_valid(token_dict: Dict[str, Any]) -> bool:
+        return all(key in token_dict for key in ['access_token', 'refresh_token', 'token_expiry'])
 
 
 if __name__ == '__main__':
